Reverted merging of destination cube and target cube
Due to faults in the path finding algorithm I have reverted my merge of
target cube and destination cube. We simply do not have time to do this.
I will further more stop revising/optimizing code from this class. I
know loads of code can be optimized. What has been produced isn't nearly
perfect, but this is because we worked in phases. If I knew what the
entire assignement was at once, I could have built the project
accordingly, unfortunately I didn't, which is why we end up with this.