Implemented PositionWorkshop
To make it easier (and not having to run over the entire world), I have
added a list of workshop positions to World. This list gets filled when
the terrain is initialized and cannot be changed after this. In
PositionWorkshop we simply use this list to loop and use the algorithm
like the one we've used before in PositionExpression.